requirements.txt
streamlit==1.30.0
torch==2.0.1
torchvision==0.15.2
Pillow==10.1.0
matplotlib==3.5.1